+++ app/src/main/java/edu/bsu/cs222/cnj/homebrewyo2/BeerBuilder.java
@@ -21,8 +21,8 @@ public abstract class BeerBuilder extends Beer {
     public abstract void buildFermentTemp(int fermentTemp);
     public abstract void buildABVPercent(double abvPercent);
     public abstract void buildIBUValue(int ibuValue);
-    public abstract void buildTargetFinalGravity(int targetFinalGravity);
-    public abstract void buildTargetOriginalGravity(int targetOriginalGravity);
+    public abstract void buildTargetFinalGravity(double targetFinalGravity);
+    public abstract void buildTargetOriginalGravity(double targetOriginalGravity);
 
     public Beer returnBeer(){
         return thisBeer;
